# Improve performance of edge iters