This PR improves the performance of the edge iterators. It also fixes a bug, where the edge iterator would not work, when `nv(g) == typemax(eltype(g))`.

The previous implementation was for `SimpleAbstractGraph`, so I kept that by using `SimpleTraits` for distinguishing between directed and undirected graphs, but all other implementations in the `SimpleGraphs` submodule seem to be only implemented for `SimpleGraph` and `SimpleDiGraph`, so we might also do that.

Some benchmarks for `collect`, it might still be useful to implement `collect` directly though (see also #1201 )

```
# sparse undirected

julia> g = dorogovtsev_mendes(10000, seed=1);

## old
julia> @btime collect(edges(g));
  342.597 μs (3 allocations: 312.59 KiB)

## new
julia> @btime collect(edges(g))
  154.607 μs (3 allocations: 312.59 KiB)


# dense undirected

julia> g = erdos_renyi(Int16(5000), 0.5, seed=1);

## old
julia> @btime collect(edges(g));
  86.561 ms (3 allocations: 23.83 MiB)

julia> @btime collect(edges(g))
  13.419 ms (3 allocations: 23.83 MiB)


# dense directed

julia> gd = SimpleDiGraph{Int16}(erdos_renyi(3500, 0.5, is_directed=true, seed=1));

## old
julia> @btime collect(edges(gd));
  26.299 ms (3 allocations: 23.36 MiB)

## new
julia> @btime collect(edges(gd))
  11.228 ms (3 allocations: 23.36 MiB)
```