@abhinavmehndiratta 2019-04-30
Why use `SimpleTraits` here ? Not every subtype of AbstractSimpleGraph is required to have a `fadj` function, so I guess we should either use `neighbors` or remove `@traitfn`.