@simonschoelly 2019-04-30
The problem with with `neighbors` is, that it does not propagate `@inbounds`. Maybe it should, but as long as it does not do that, we would have an additional bound check in every iteration.