@sbromberger 2019-05-01
@simonschoelly is this ready to go, or is there still an issue with overloading an unexported function from `Base`?