@simonschoelly 2019-05-01
I think you are confusing this PR with #1137 - we are not overloading any unexported functions here.
@abhinavmehndiratta had some objections, so I don't know if I still should do some changes here - otherwise we might merge this.