@sbromberger 2019-05-01
> I think you are confusing this PR with #1137

I am. Thanks for setting me straight.

I'm generally ok with adding `SimpleTraits` here since it's a good use of the function, but I note that you could probably pull the inner loops out into their own functions and use `is_directed()` to call whichever one works. This might have a performance impact but I can't imagine it'd be too big.