@simonschoelly 2019-05-15
So, should I write a version using `is_directed` or will we merge it like this?